# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ethnos.app Academic Bibliography API",
    "version": "2.0.0",
    "description": "\nREST API for academic bibliography: search, works, persons, institutions, venues, courses, citations, collaborations and health.\n\nHighlights\n- Public endpoints; selected admin/metrics routes require 'x-access-key'.\n- Standard response envelope: { status, data, pagination?, meta? }.\n- Pagination supports page/limit and offset/limit.\n- Identifiers: DOI, ORCID, ROR, ISSN.\n\nNotes\n- Rate limiting is configurable via environment and enabled by default.\n- Performance and dataset size vary by deployment.\n        ",
    "contact": {
      "name": "Bruno Cesar Cunha Cruz, PhD Student"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-developer-orcid": "0000-0001-8652-2333",
    "x-institution": "PPGAS/MN/UFRJ"
  },
  "servers": [
    {
      "url": "https://api.ethnos.app",
      "description": "Production API Server - Enterprise-ready academic research infrastructure"
    },
    {
      "url": "http://localhost:1210",
      "description": "Development Server - Local testing and development environment"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "XAccessKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-key",
        "description": "Internal access key required for protected endpoints (security, dashboard, health except /liveness)"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "Response payload. Varies per endpoint."
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "status",
          "data"
        ]
      },
...
      "Error": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Resource not found"
          },
          "code": {
            "type": "string",
            "example": "NOT_FOUND"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Venue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "American Anthropologist",
            "description": "Official name of the venue"
          },
          "type": {
            "type": "string",
            "enum": [
              "JOURNAL",
              "CONFERENCE",
              "REPOSITORY",
              "BOOK_SERIES"
            ],
            "example": "JOURNAL"
          },
          "impact_factor": {
            "type": "number",
            "format": "float",
            "nullable": true,
            "example": 3.214
          },
          "citescore": {
            "type": "number",
            "format": "float",
            "nullable": true,
            "example": 5.6
          },
          "sjr": {
            "type": "number",
            "format": "float",
            "nullable": true,
            "example": 1.12
          },
          "snip": {
            "type": "number",
            "format": "float",
            "nullable": true,
            "example": 0.98
          },
          "open_access": {
            "type": "boolean",
            "nullable": true,
            "description": "Indicates whether the venue is fully open access",
            "example": false
          },
          "aggregation_type": {
            "type": "string",
            "nullable": true,
            "example": "journal"
          },
          "coverage_start_year": {
            "type": "integer",
            "nullable": true,
            "example": 1984
          },
          "coverage_end_year": {
            "type": "integer",
            "nullable": true,
            "example": 2025
          },
          "works_count": {
            "type": "integer",
            "example": 1487
          },
          "homepage_url": {
            "type": "string",
            "nullable": true,
            "example": "http://www.jps.auckland.ac.nz"
          },
          "country_code": {
            "type": "string",
            "nullable": true,
            "example": "NZ"
          },
          "is_in_doaj": {
            "type": "boolean",
            "nullable": true,
            "example": false
          },
          "publications_count": {
            "type": "integer",
            "example": 1523
          },
          "last_validated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2025-03-12T10:15:00Z"
          },
          "validation_status": {
            "type": "string",
            "enum": [
              "PENDING",
              "VALIDATED",
              "NOT_FOUND",
              "FAILED"
            ],
            "example": "VALIDATED"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "1998-01-15T00:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-02-01T08:22:00Z"
          },
          "identifiers": {
            "type": "object",
            "properties": {
              "issn": {
                "type": "string",
                "nullable": true,
                "example": "0002-7294"
              },
              "eissn": {
                "type": "string",
                "nullable": true,
                "example": "1548-1433"
              },
              "scopus_source_id": {
                "type": "string",
                "nullable": true,
                "example": "12345"
              },
              "external": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "example": {
                  "WIKIDATA": "https://www.wikidata.org/entity/Q15760156",
                  "OPENALEX": "https://openalex.org/S189780351"
                }
              }
            }
          },
          "publisher": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "nullable": true,
                "example": 42
              },
              "name": {
                "type": "string",
                "nullable": true,
                "example": "Wiley-Blackwell"
              },
              "type": {
                "type": "string",
                "nullable": true,
                "example": "PUBLISHER"
              },
              "country_code": {
                "type": "string",
                "nullable": true,
                "example": "US"
              }
            }
          },
          "metrics": {
            "$ref": "#/components/schemas/VenueMetrics"
          },
          "subjects": {
            "type": "array",
            "description": "Subjects associated with the venue (list responses limit to top 5 by default)",
            "items": {
              "type": "object",
              "properties": {
                "subject_id": {
                  "type": "integer",
                  "nullable": true,
                  "example": 1205
                },
                "term": {
                  "type": "string",
                  "example": "Anthropology"
                },
                "score": {
                  "type": "number",
                  "format": "float",
                  "nullable": true,
                  "example": 0.92
                }
              }
            }
          },
          "terms": {
            "type": "array",
            "description": "Ordered list of subject terms extracted for the venue",
            "items": {
              "type": "string",
              "example": "Anthropology"
            }
          },
          "keywords": {
            "type": "array",
            "description": "Deduplicated subject terms normalized for keyword usage",
            "items": {
              "type": "string",
              "example": "anthropology"
            }
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 6894,
            "description": "Total number of results"
          },
          "page": {
            "type": "integer",
            "example": 1,
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "example": 20,
            "description": "Number of results per page"
          },
          "totalPages": {
            "type": "integer",
            "example": 345,
            "description": "Total number of pages"
          },
          "hasNext": {
            "type": "boolean",
            "example": true,
            "description": "Whether there is a next page"
          },
          "hasPrev": {
            "type": "boolean",
            "example": false,
            "description": "Whether there is a previous page"
          }
        }
      },
...
      "VenueMetrics": {
        "type": "object",
        "properties": {
          "publications_count": {
            "type": "integer",
            "example": 1523
          },
          "works_count": {
            "type": "integer",
            "example": 1487
          },
          "unique_authors": {
            "type": "integer",
            "example": 512
          },
          "cited_by_count": {
            "type": "integer",
            "example": 4375
          },
          "h_index": {
            "type": "integer",
            "example": 32
          },
          "i10_index": {
            "type": "integer",
            "example": 144
          },
          "open_access_publications": {
            "type": "integer",
            "example": 420
          },
          "open_access_percentage": {
            "type": "number",
            "format": "float",
            "nullable": true,
            "example": 27.53
          },
          "total_citations": {
            "type": "integer",
            "example": 18345
          },
          "avg_citations": {
            "type": "number",
            "format": "float",
            "nullable": true,
            "example": 12.04
          },
          "total_downloads": {
            "type": "integer",
            "example": 92341
          },
          "first_publication_year": {
            "type": "integer",
            "nullable": true,
            "example": 1984
          },
          "latest_publication_year": {
            "type": "integer",
            "nullable": true,
            "example": 2025
          }
        }
      },
...
    },
...
    "responses": {
      "Success": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SuccessEnvelope"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request - Invalid input parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Invalid venue ID",
              "code": "VALIDATION_ERROR"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Resource not found",
              "code": "NOT_FOUND"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "status": "error",
              "message": "Internal server error",
              "code": "INTERNAL_ERROR"
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/venues": {
      "get": {
        "summary": "Get list of academic venues",
        "description": "Retrieve a paginated list of academic venues including journals, conferences, repositories, and book series. Supports filtering and sorting.",
        "tags": [
          "Venues"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of results per page (max 100)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Number of items to skip (alternative to page parameter)"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "JOURNAL",
                "CONFERENCE",
                "REPOSITORY",
                "BOOK_SERIES"
              ]
            },
            "description": "Filter by venue type",
            "example": "JOURNAL"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "description": "Search term to filter venues by name",
            "example": "Nature"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "name",
                "type",
                "impact_factor",
                "works_count",
                "id"
              ]
            },
            "description": "Field to sort by",
            "example": "works_count"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            },
            "description": "Sort order",
            "example": "DESC"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/venues/statistics": {
      "get": {
        "summary": "Get venue statistics",
        "description": "Retrieve comprehensive statistics about venues in the system including type distribution and metrics",
        "tags": [
          "Venues"
        ],
        "responses": {
          "200": {
            "description": "Venue statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total_venues": {
                          "type": "integer",
                          "example": 1563
                        },
                        "by_type": {
                          "type": "object",
                          "properties": {
                            "JOURNAL": {
                              "type": "integer",
                              "example": 892
                            },
                            "CONFERENCE": {
                              "type": "integer",
                              "example": 445
                            },
                            "REPOSITORY": {
                              "type": "integer",
                              "example": 156
                            },
                            "BOOK_SERIES": {
                              "type": "integer",
                              "example": 70
                            }
                          }
                        },
                        "avg_impact_factor": {
                          "type": "number",
                          "format": "float",
                          "example": 3.2
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/venues/search": {
      "get": {
        "summary": "Search venues",
        "description": "Search for venues by name and filter by type",
        "tags": [
          "Venues"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            },
            "description": "Search query for venue name",
            "example": "Nature"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "JOURNAL",
                "CONFERENCE",
                "REPOSITORY",
                "BOOK_SERIES"
              ]
            },
            "description": "Filter by venue type"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of results per page"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Venue"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/venues/{id}": {
      "get": {
        "summary": "Get venue by ID",
        "description": "Retrieve detailed information about a specific venue",
        "tags": [
          "Venues"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Venue ID",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Venue details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Venue"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/venues/{id}/works": {
      "get": {
        "summary": "Get works published in a venue with citations and authors",
        "description": "Retrieve all works published in a specific venue ordered by citation count and publication year, including complete author information",
        "tags": [
          "Venues"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Venue ID",
            "example": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Number of results per page"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of items to skip"
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer",
              "minimum": 1900,
              "maximum": 2100
            },
            "description": "Filter by publication year"
          }
        ],
        "responses": {
          "200": {
            "description": "Works in venue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Work"
                      }
                    },
                    "pagination": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}